requirements.txt
aiohttp==3.7.4.post0
astroid==2.9.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
coverage==6.3.2
discord.py==1.7.3
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.8.0
mccabe==0.6.1
multidict==6.0.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.8
wrapt==1.13.3
yarl==1.7.2
